# txscript: Decouple and optimize sighash calc.