… txscript/data/many_inputs_tx.hex
… txscript/sighash.go