**This PR builds on top of #1178 and #1181**.

This PR consists of 2 commits.  The first commit simply adds a benchmark for signature hash calculation using a transaction with many inputs while the second one contains the meat of the changes.

This modifies the signature hash calculation logic to completely decouple it from the wire transaction serialization and significantly optimizes it in the process.  It also significantly improves the comments to specifically call out the semantics.

This change is highly desirable because it has several beneficial properties:

- The signature hash semantics are much clearer and specific
- There is no longer a need to copy the entire transaction and modify the relevant portions since the necessary substitutions are made on the fly
- Allows much faster calculation by serializing directly into byte slices and avoiding all of the additional error handling logic dedicated to handling network streams
- Provides the possibility of changing the wire format without breaking signature hash calculation
  - Note that the caching portion still relies on the wire format, but that can be addressed in future commits since it is merely an optimization that can also be decoupled

The following is a before and after comparison of signature hash calculation for both speed and memory allocations for a transaction with many inputs:

```
benchmark                old ns/op     new ns/op     delta
------------------------------------------------------------
BenchmarkCalcSigHash     6299714       1551740       -75.37%

benchmark                old allocs    new allocs    delta
------------------------------------------------------------
BenchmarkCalcSigHash     18601         1691          -90.91%
```