@arajpoot 2018-05-02
Will it help bottom line in more uses and partnership.